encripted passworld

+++ src/controller/model/response/user_response.go
@@ -4,5 +4,5 @@ type UserResponse struct {
 	ID    string `json:"id"`
 	Email string `json:"email"`
 	Name  string `json:"name"`
-	Age   int8   `json:"age"`
+	Age   int16  `json:"age"`
 }
